iiod: Create default context instead of local
Instead of creating a local context unconditionally, create the default
context. This allows overriding the context used within IIOD via the
IIOD_REMOTE variable, and is very useful for debugging.

This should be pretty safe, as the IIOD_REMOTE variable is very unlikely
to be set on the platforms where IIOD is supposed to run.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>