stream: Simplify iio_stream_get_next_block()
The bytes_used field can now be set to 0 to default to the block size.

Note that for RX transfers, iio_block_enqueue() was already called
with bytes_used == 0, which previously meant "transmit 0 bytes and
queue for receive". Because the "bytes_used" field is now handled as
for RX buffers as well, the behaviour is now "queue for receiving a
block size worth of data" which is functionally the same.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>