local: Update to new DMABUF API
The first version of the DMABUF API was refused in the Linux kernel.

Update the local backend to use the new version of the DMABUF API, which
will hopefully be merged this time.

The main difference between the two interfaces is that the IIO subsystem
does not create the DMABUFs anymore; it only supports importing them,
drastically reducing the complexity of the kernel code.

This means that the DMABUFs have to be created by something else. The
local backend now uses the 'udmabuf' kernel driver to create DMABUF
objects in system memory from 'memfd' objects.

Signed-off-by: Paul Cercueil <paul@crapouillou.net>